requirements.txt
Django
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-decouple
gunicorn
drf_yasg
whitenoise